apps/data-worker/pyproject.toml
[project]
name = "data-worker"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Gustavo Henrique", email = "ghfc2000@hotmail.com"},
]
dependencies = [
    "biopython>=1.85",
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "python src/main.py"
